Starting to fix some of the unit tests.  More work to do though
