+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+markers =
+    coreEnvReddit
+    slow: marks tests as slow (deselect with '-m "not slow"')
+    serial
