+++ tests/worker/outputs/test_discord.py
